Added support for Xiao ESP32-S3 mounted on my custom carrier board. This includes a new partition table, updated deviceconfig, and some minor changes to the web installer to support the new board.
Added support for four led strips running
different effects on each strip. This includes an
updated web server to allow selection of effects
for each strip, and updated effectmanager to
support multiple strips.

+++ include/effects/matrix/PatternRadar.h
@@ -79,7 +79,7 @@ public:
       CRGB color = graphics.ColorFromCurrentPalette(hue);
       uint8_t x = graphics.mapcos8(theta, offset, (MATRIX_WIDTH - 1) - offset);
       uint8_t y = graphics.mapsin8(theta, offset, (MATRIX_HEIGHT - 1) - offset);
-      size_t xzy = graphics.xy(x, y);
+      uint16_t xzy = graphics.xy(x, y);
       graphics.leds[xzy] = color;
 
       EVERY_N_MILLIS(25)
